software/RIOT*: greatly improved font rendering
This adds a new font from the https://github.com/trip5/Matrix-Fonts
repository and a new binary font encoding that better fits
fonts with variable width.

The led_matrix module has been updated to allow specifying the
font used to render text and the API has been improved.

A simple heuristic to detect when adding a blank column between
to glyphs is needed for readability, or when it should be
omitted (e.g. between "o" and "T") because there already is one
empty pixel between the glyphs anyway and it would look ugly
if the space is added.

In addition, the led_matrix_fb_switch API has changed: A frame
number now is specified when the scratch frame buffer should
become active. This prevents switching the frame buffer in
the middle of drawing a frame and should improve readability.